Formularios WS y vista convenios SD
Desde la bitácora se puede asignar WS a los convenios y se crea página para usuarios de SD.
