+++ convenios_app/informes/routes.py
@@ -3,7 +3,6 @@ from flask import render_template, request, Blueprint, url_for, redirect, flash,
 from flask_login import current_user, login_required
 from convenios_app.models import (Institucion, Equipo, Persona, Convenio, SdInvolucrada, BitacoraAnalista,
                                   BitacoraTarea, TrayectoriaEtapa, TrayectoriaEquipo)
-from convenios_app.informes.forms import MisConveniosInfoConvenioForm, MisConveniosBitacoraForm, MisConveniosTareaForm
 from convenios_app import db
 from sqlalchemy import and_, or_
 from convenios_app.bitacoras.utils import actualizar_trayectoria_equipo, actualizar_convenio, obtener_iniciales
@@ -392,168 +391,14 @@ scheduler.start()
 estadisticas_convenios()
 
 
-@informes.route('/mis_convenios/<int:id_persona>', methods=['GET', 'POST'])
-@login_required
-@analista_only
-def mis_convenios(id_persona):
-    if current_user.permisos != 'Admin':
-        persona = Persona.query.get_or_404(id_persona)
-        if persona.id != current_user.id_persona:
-            abort(403)
-
-    convenios_analista_query = Convenio.query.filter(Convenio.id_coord_sii == id_persona).all()
-
-    # Tareas pendientes
-    id_convenios = [convenio.id for convenio in convenios_analista_query]
-    tareas_query = BitacoraTarea.query.filter(and_(BitacoraTarea.id_convenio.in_(id_convenios)),
-                                              BitacoraTarea.estado == 'Pendiente').order_by(
-        BitacoraTarea.plazo.asc()).all()
-    tareas_pendientes = []
-    for tarea in tareas_query:
-        tareas_pendientes.append({
-            'id_convenio': tarea.convenio.id,
-            'nombre_convenio': generar_nombre_convenio(tarea.convenio),
-            'tarea': tarea.tarea,
-            'plazo': tarea.plazo,
-            'id_tarea': tarea.id
-        })
-    tareas_pendientes.sort(key=lambda dict: dict['plazo'])
-
-    # Actualizar bitácora
-    convenios_select = [(convenio.id, generar_nombre_convenio(convenio)) for convenio in convenios_analista_query]
-    convenios_select.insert(0, (0, 'Seleccionar convenio'))
-
-    # Formulario información de convenio
-    form_info_convenio = MisConveniosInfoConvenioForm()
-    form_info_convenio.etapa.choices.insert(0, (0, 'Seleccione convenio para ver'))
-    if 'informacion_convenio' in request.form and form_info_convenio.validate_on_submit():
-        convenio = Convenio.query.get(form_info_convenio.id_convenio.data)
-        # Actualizar áreas y etapas
-        form_info_equipos = [
-            (form_info_convenio.id_trayectoriaEquipo_1.data, form_info_convenio.equipo_1.data, form_info_convenio.fecha_equipo_1.data),
-            (form_info_convenio.id_trayectoriaEquipo_2.data, form_info_convenio.equipo_2.data, form_info_convenio.fecha_equipo_2.data),
-            (form_info_convenio.id_trayectoriaEquipo_3.data, form_info_convenio.equipo_3.data, form_info_convenio.fecha_equipo_3.data),
-            (form_info_convenio.id_trayectoriaEquipo_4.data, form_info_convenio.equipo_4.data, form_info_convenio.fecha_equipo_4.data)
-        ]
-
-        # Trayectoria etapas
-        etapa_actual = TrayectoriaEtapa.query.get(form_info_convenio.id_trayectoriaEtapa.data)
-        if etapa_actual.id_etapa != int(form_info_convenio.etapa.data):
-            etapa_actual.actualizar_trayectoria_etapa(form_info_convenio)
-            # Si se finaliza el proceso
-            if form_info_convenio.etapa.data == '5':
-                # Finalizar etapa 'Finalizado'
-                ultima_etapa = TrayectoriaEtapa.query.filter(
-                    and_(TrayectoriaEtapa.id_convenio == form_info_convenio.id_convenio.data,
-                         TrayectoriaEtapa.salida == None)).first()
-                ultima_etapa.salida = form_info_convenio.fecha_etapa.data
-                ultima_etapa.timestamp_salida = datetime.today()
-                # Desasignar equipos de trabajo
-                for equipo in form_info_equipos:
-                    actualizar_trayectoria_equipo(equipo[0], '0', form_info_convenio.fecha_etapa.data, form_info_convenio.id_convenio.data)
-                # Eliminar tareas pendientes
-                tareas_pendientes_convenio = BitacoraTarea.query.filter(
-                    and_(BitacoraTarea.id_convenio == form_info_convenio.id_convenio.data,
-                         BitacoraTarea.estado == 'Pendiente')).all()
-                for tarea in tareas_pendientes_convenio:
-                    tarea.estado = 'Eliminado'
-                    tarea.timestamp = datetime.today()
-                # Dejar registro en la bitácora
-                ultimo_registro_bitacora = BitacoraAnalista(
-                    observacion='Se cambia etapa a: Finalizado',
-                    fecha=form_info_convenio.fecha_etapa.data,
-                    timestamp=datetime.today(),
-                    id_convenio=form_info_convenio.id_convenio.data,
-                    id_autor=current_user.id
-                )
-                db.session.add(ultimo_registro_bitacora)
-                db.session.commit()
-
-                flash(
-                    f'{generar_nombre_convenio(convenio)} ha sido finalizado. No olvide cambiar el estado del {convenio.tipo}.',
-                    'warning')
-                return redirect(url_for('bitacoras.editar_convenio', id_convenio=form_info_convenio.id_convenio.data))
-            # Si convenio sigue en proceso
-            else:
-                # Dejar registro en bitácora del analista
-                observacion_cambio_etapa = BitacoraAnalista(
-                    observacion=f'Se cambia etapa a: {dict(ETAPAS).get(int(form_info_convenio.etapa.data))}',
-                    fecha=form_info_convenio.fecha_etapa.data,
-                    timestamp=datetime.today(),
-                    id_convenio=form_info_convenio.id_convenio.data,
-                    id_autor=current_user.id
-                )
-                db.session.add(observacion_cambio_etapa)
-                db.session.commit()
-
-        # Trayectoria equipos
-        for equipo in form_info_equipos:
-            actualizar_trayectoria_equipo(equipo[0], equipo[1], equipo[2], form_info_convenio.id_convenio.data)
-        flash(f'Se ha actualizado la información de {generar_nombre_convenio(convenio)}', 'success')
-        return redirect(url_for('informes.mis_convenios', id_persona=id_persona))
-
-    # Formulario nueva observación bitácora
-    form_bitacora = MisConveniosBitacoraForm()
-    if 'bitacora_analista' in request.form and form_bitacora.validate_on_submit():
-        # Agregar nueva observación
-        nueva_observacion = BitacoraAnalista(
-            observacion=form_bitacora.observacion.data,
-            fecha=form_bitacora.fecha.data,
-            timestamp=datetime.today(),
-            id_convenio=form_bitacora.id_convenio_bitacora.data,
-            id_autor=current_user.id
-        )
-        db.session.add(nueva_observacion)
-        db.session.commit()
-        convenio = Convenio.query.get(form_bitacora.id_convenio_bitacora.data)
-        flash(f'Se actualizado la bitácora de {generar_nombre_convenio(convenio)}', 'success')
-        return redirect(url_for('informes.mis_convenios', id_persona=id_persona))
-
-    # Formulario nueva tarea
-    form_tarea = MisConveniosTareaForm()
-    if 'nueva_tarea' in request.form and form_tarea.validate_on_submit():
-        # Agregar nueva tarea
-        nueva_tarea = BitacoraTarea(
-            tarea=form_tarea.tarea.data,
-            plazo=form_tarea.plazo.data,
-            timestamp=datetime.today(),
-            id_convenio=form_tarea.id_convenio_tarea.data,
-            id_autor=current_user.id
-        )
-        db.session.add(nueva_tarea)
-        db.session.commit()
-        convenio = Convenio.query.get(form_tarea.id_convenio_tarea.data)
-        flash(f'Se ha agregado tarea a {generar_nombre_convenio(convenio)}', 'success')
-        return redirect(url_for('informes.mis_convenios', id_persona=id_persona))
+@informes.route('/otros_convenios')
+def otros_convenios():
+    return render_template('informes/otros_convenios.html')
 
 
-    # Estado actual de mis convenios
-    tabla_estado_actual = []
-    for convenio in convenios_analista_query:
-        nombre = generar_nombre_convenio(convenio)
-        observacion_query = BitacoraAnalista.query.filter(and_(BitacoraAnalista.id_convenio == convenio.id,
-                                                               BitacoraAnalista.estado != 'Eliminado')).order_by(
-            BitacoraAnalista.fecha.desc(), BitacoraAnalista.timestamp.desc()).first()
-        ultima_observacion = f'({datetime.strftime(observacion_query.fecha, "%d-%m-%Y")}) {observacion_query.observacion}'
-        etapa = convenio.estado
-        suplente = f'<p align="center">{(lambda sup: obtener_iniciales(sup.nombre) if sup else "")(convenio.sup_sii)}</p>'
-        link_resolucion = (lambda
-                               link: f'<a target="_blank" class="text-center" style="text-decoration: none; color: #000;" href="{link}">'
-                                     f'<i class="fas fa-eye pt-2 text-center btn-lg"></i></a>' if link else
-        '<div class="text-center"><i class="fas fa-eye-slash pt-2 text-center text-muted btn-lg"></i></div>')(
-            convenio.link_resolucion)
-        tabla_estado_actual.append([nombre, etapa, ultima_observacion, suplente, link_resolucion, convenio.id])
-    # Ordenar tabla
-    tabla_estado_actual.sort(key=lambda lista: lista[0])
-    # Agregar link al nombre del convenio y botar el id
-    for convenio in tabla_estado_actual:
-        convenio[0] = f'<a style="text-decoration: none; color: #000;" href={url_for("bitacoras.bitacora_convenio", id_convenio=convenio[5])}>' \
-                 f'{convenio[0]} <i class="fa-solid fa-keyboard fa-fw"></i></a>'
-        convenio.pop()
-
-    return render_template('informes/mis_convenios.html', tareas_pendientes=tareas_pendientes, hoy=date.today(),
-                           id_persona=id_persona, convenios_select=convenios_select, form_info=form_info_convenio,
-                           form_bitacora=form_bitacora, form_tarea=form_tarea, tabla_estado_actual=tabla_estado_actual)
+@informes.route('/detalle_otros_convenios/<int:id_convenio>', methods=['GET', 'POST'])
+def detalle_otros_convenios(id_convenio):
+    return 'detalle'
 
 
 @informes.route('/convenios_en_proceso')
@@ -1205,7 +1050,7 @@ def convenios_en_produccion():
         'mes_firma': int(datetime.strftime(convenio.fecha_documento, '%m')) - 1,
         'año_firma': datetime.strftime(convenio.fecha_documento, '%Y'),
         'resolucion': (lambda
-                           convenio: f'{convenio.nro_resolucion} del {datetime.strftime(convenio.fecha_resolucion, "%d-%m-%Y")}' if convenio.fecha_resolucion != None else 'Sin resolución')(
+                           convenio: f'N°{convenio.nro_resolucion} del {datetime.strftime(convenio.fecha_resolucion, "%d-%m-%Y")}' if convenio.fecha_resolucion != None else 'Sin resolución')(
             convenio),
         'link_resolucion': (lambda
                                 link: f'<a target="_blank" class="text-center" style="text-decoration: none; color: #000;" href="{link}">'
@@ -1579,152 +1424,46 @@ def detalle_convenio_en_produccion(id_convenio):
                            tareas_equipos=tareas_equipos)
 
 
-@informes.route('/obtener_info_convenio/<int:id_convenio>')
-def obtener_info_convenio(id_convenio):
-    convenio_query = Convenio.query.get(id_convenio)
-    etapa_query = (lambda estado: TrayectoriaEtapa.query.filter(and_(TrayectoriaEtapa.id_convenio == id_convenio,
-                                                         TrayectoriaEtapa.salida == None)).first() if estado == 'En proceso' else
-                    TrayectoriaEtapa.query.filter(and_(TrayectoriaEtapa.id_convenio == id_convenio,
-                                       TrayectoriaEtapa.id_etapa == 5)).order_by(TrayectoriaEtapa.ingreso.asc()).first()
-                   )(convenio_query.estado)
-    # Calcular días en proceso
-    dias_proceso = 0
-    for trayectoEtapa in TrayectoriaEtapa.query.filter(TrayectoriaEtapa.id_convenio == id_convenio).all():
-        salida_etapa = (lambda salida: salida if salida != None else date.today())(trayectoEtapa.salida)
-        dias_proceso += (salida_etapa - trayectoEtapa.ingreso).days
-
-    # Calcular áreas actuales y días en proceso
-    equipos_query = TrayectoriaEquipo.query.filter(
-        and_(TrayectoriaEquipo.id_convenio == id_convenio, TrayectoriaEquipo.salida == None)).order_by(
-        TrayectoriaEquipo.ingreso.asc()).all()
-    equipos = []
-    for trayectoEquipo in equipos_query:
-        equipos.append({
-            'id_trayectoEquipo': trayectoEquipo.id,
-            'id_equipo': trayectoEquipo.id_equipo,
-            'ingreso': datetime.strftime(trayectoEquipo.ingreso, '%Y-%m-%d'),
-            'dias_equipo': (date.today() - trayectoEquipo.ingreso).days
-        })
-    for i in range(4 - len(equipos_query)):
-        equipos.append({
-            'id_trayectoEquipo': 0,
-            'id_equipo': 0,
-            'ingreso': "",
-            'dias_equipo': ""
-        })
-
-    info_convenio = {
-        'id_convenio': id_convenio,
-        'dias_proceso': dias_proceso,
-        'etapa': {
-            'id_trayectoEtapa': etapa_query.id,
-            'id_etapa': etapa_query.id_etapa,
-            'fecha_etapa': datetime.strftime(etapa_query.ingreso, '%Y-%m-%d'),
-            'dias_etapa': (lambda etapa: (date.today() - etapa_query.ingreso).days if etapa != 5 else '')(etapa_query.id_etapa)
-        },
-        'equipos': equipos
-
-    }
-    return jsonify(info_convenio)
 
 
 
 
 
 
-def obtener_info_convenios_en_proceso():
-    convenios_en_proceso = Convenio.query.filter(Convenio.estado == 'En proceso').all()
-    info_convenios = []
-    for convenio in convenios_en_proceso:
-        ultima_observacion = BitacoraAnalista.query.filter(and_(BitacoraAnalista.id_convenio == convenio.id,
-                                                                BitacoraAnalista.estado != 'Eliminado')).order_by(
-            BitacoraAnalista.fecha.desc(), BitacoraAnalista.timestamp.desc()).first()
-        proxima_tarea = BitacoraTarea.query.filter(
-            and_(BitacoraTarea.id_convenio == convenio.id, BitacoraTarea.estado == 'Pendiente')).order_by(
-            BitacoraTarea.plazo.asc(), BitacoraTarea.timestamp.asc()).first()
-        etapa_actual = TrayectoriaEtapa.query.filter(
-            and_(TrayectoriaEtapa.id_convenio == convenio.id, TrayectoriaEtapa.salida == None)).first()
-
-        equipos_actual = TrayectoriaEquipo.query.filter(
-            and_(TrayectoriaEquipo.id_convenio == convenio.id, TrayectoriaEquipo.salida == None)).order_by(
-            TrayectoriaEquipo.ingreso.asc()).all()
-        equipos = []
-        for equipo in equipos_actual:
-            equipos.append(f'{equipo.equipo.sigla} ({(date.today() - equipo.ingreso).days})')
-        for _ in range(4 - len(equipos_actual)):
-            equipos.append("")
-
-        info_convenios.append({
-            'nombre': generar_nombre_convenio(convenio),
-            # 'encargados': f'{obtener_iniciales(convenio.coord_sii.nombre)}  )',   #({obtener_iniciales(convenio.sup_sii.nombre)})',
-            # 'coordinador': convenio.coord_sii.nombre,
-            # 'suplente': convenio.sup_sii.nombre,
-            # coord_ie, sup_ie
-            'observacion': f'{ultima_observacion.fecha}: {ultima_observacion.observacion}',
-            'tarea': (lambda tarea: f'{tarea.plazo}: {tarea.tarea}' if tarea != None else "Sin tareas pendientes")(
-                proxima_tarea),
-            'plazo_tarea': (lambda tarea: tarea.plazo if tarea != None else "Sin tareas pendientes")(proxima_tarea),
-            'etapa': f'{etapa_actual.etapa.etapa} ({(date.today() - etapa_actual.ingreso).days})',
-            'equipos': equipos,
-        })
-    info_convenios.sort(key=lambda dict: dict['nombre'])
-    return info_convenios
-
-
-@informes.route('/resumen_convenios_en_proceso')
-def resumen_convenios_en_proceso():
-    info_convenios = obtener_info_convenios_en_proceso()
-    return render_template('informes/convenios_en_proceso.html', info_convenios=info_convenios, hoy=date.today())
-
-
+# def obtener_info_convenios_en_proceso():
+#     convenios_en_proceso = Convenio.query.filter(Convenio.estado == 'En proceso').all()
+#     info_convenios = []
+#     for convenio in convenios_en_proceso:
+#         ultima_observacion = BitacoraAnalista.query.filter(and_(BitacoraAnalista.id_convenio == convenio.id,
+#                                                                 BitacoraAnalista.estado != 'Eliminado')).order_by(
+#             BitacoraAnalista.fecha.desc(), BitacoraAnalista.timestamp.desc()).first()
+#         proxima_tarea = BitacoraTarea.query.filter(
+#             and_(BitacoraTarea.id_convenio == convenio.id, BitacoraTarea.estado == 'Pendiente')).order_by(
+#             BitacoraTarea.plazo.asc(), BitacoraTarea.timestamp.asc()).first()
+#         etapa_actual = TrayectoriaEtapa.query.filter(
+#             and_(TrayectoriaEtapa.id_convenio == convenio.id, TrayectoriaEtapa.salida == None)).first()
 #
-# @informes.route('/resumen_convenios_en_proceso_pdf')
-# def resumen_convenios_en_proceso_pdf():
-#     info_convenios = obtener_info_convenios_en_proceso()
-#     informe_rendered = render_template('/informes/convenios_en_proceso_pdf.html', info_convenios=info_convenios, hoy=date.today())
-#     informe_rendered.encode('utf-8')
-#     try:
-#         informe_pdf = pdfkit.from_string(informe_rendered, 'C:\Convenios_App\convenios_app\documentos\informe_en_proceso.pdf', configuration=config)
-#     except IOError:
-#         pass
+#         equipos_actual = TrayectoriaEquipo.query.filter(
+#             and_(TrayectoriaEquipo.id_convenio == convenio.id, TrayectoriaEquipo.salida == None)).order_by(
+#             TrayectoriaEquipo.ingreso.asc()).all()
+#         equipos = []
+#         for equipo in equipos_actual:
+#             equipos.append(f'{equipo.equipo.sigla} ({(date.today() - equipo.ingreso).days})')
+#         for _ in range(4 - len(equipos_actual)):
+#             equipos.append("")
 #
-#     return send_from_directory('C:\Convenios_App\convenios_app\documentos', 'informe_en_proceso.pdf', as_attachment=True, attachment_filename=f'Convenios_en_proceso_{date.today()}.pdf')
-
-
-@informes.route('/informe_aiet')
-def informe_aiet():
-    info_convenios = obtener_info_convenios_en_proceso()
-    return render_template('informes/informe_aiet.html', info_convenios=info_convenios, hoy=date.today())
-
-
-#
-# @informes.route('/informe_aiet_pdf')
-# def informe_aiet_pdf():
-#     info_convenios = obtener_info_convenios_en_proceso()
-#     informe_rendered = render_template('/informes/informe_aiet_pdf.html', info_convenios=info_convenios, hoy=date.today())
-#     try:
-#         informe_pdf = pdfkit.from_string(informe_rendered, 'C:\Convenios_App\convenios_app\documentos\informe_aiet.pdf', configuration=config )
-#     except IOError:
-#         pass
-#
-#     return send_from_directory('C:\Convenios_App\convenios_app\documentos', 'informe_aiet.pdf', as_attachment=True, attachment_filename=f'Informe_aiet_{date.today()}.pdf')
-
-
-@informes.route('/mis_tareas/<int:id_persona>')
-def mis_tareas(id_persona):
-    # Select persona
-    analistas = [(persona.id, persona.nombre) for persona in
-                 Persona.query.filter(Persona.id_equipo == 1).order_by(Persona.nombre.asc()).all()]
-
-    # Obtener convenios en proceso del analista
-    convenios_analista = [convenio.id for convenio in Convenio.query.filter(
-        and_(Convenio.id_coord_sii == id_persona, Convenio.estado == 'En proceso')).all()]
-    tareas_query = BitacoraTarea.query.filter(and_(BitacoraTarea.id_convenio.in_(convenios_analista)),
-                                              BitacoraTarea.estado == 'Pendiente').order_by(
-        BitacoraTarea.plazo.asc()).all()
-    tareas = [(tarea.tarea, tarea.plazo, generar_nombre_convenio(tarea.convenio), tarea.id, tarea.id_convenio) for tarea
-              in tareas_query]
-    tareas.sort(key=lambda tup: tup[2])
-
-    return render_template('informes/mis_tareas.html', analistas=analistas, tareas=tareas, id_persona=id_persona,
-                           hoy=date.today())
+#         info_convenios.append({
+#             'nombre': generar_nombre_convenio(convenio),
+#             # 'encargados': f'{obtener_iniciales(convenio.coord_sii.nombre)}  )',   #({obtener_iniciales(convenio.sup_sii.nombre)})',
+#             # 'coordinador': convenio.coord_sii.nombre,
+#             # 'suplente': convenio.sup_sii.nombre,
+#             # coord_ie, sup_ie
+#             'observacion': f'{ultima_observacion.fecha}: {ultima_observacion.observacion}',
+#             'tarea': (lambda tarea: f'{tarea.plazo}: {tarea.tarea}' if tarea != None else "Sin tareas pendientes")(
+#                 proxima_tarea),
+#             'plazo_tarea': (lambda tarea: tarea.plazo if tarea != None else "Sin tareas pendientes")(proxima_tarea),
+#             'etapa': f'{etapa_actual.etapa.etapa} ({(date.today() - etapa_actual.ingreso).days})',
+#             'equipos': equipos,
+#         })
+#     info_convenios.sort(key=lambda dict: dict['nombre'])
+#     return info_convenios
