+++ convenios_app/static/css/styles.css
@@ -1,3 +1,11 @@
+.green {
+    color: #62BD69;
+}
+
+.gray {
+  color: #7F7F7F;
+}
+
 /* Sidebar */
 #brand-link {
   color: #FFF;
