+++ convenios_app/templates/bitacoras/bitacora.html
@@ -53,10 +53,10 @@
 </div>
 
 <!-- CONVENIOS EN PROCESO -->
-<legend class="border-bottom mb-3">Convenios en producción
+<legend class="border-bottom mb-3">
   <a style="text-decoration: none; color: #000;" data-bs-toggle="collapse"
   href="#conveniosEnProduccion" aria-expanded="false" aria-controls="conveniosEnProduccion">
-  <i class="fas fa-angle-down"></i></a></legend>
+  Convenios en producción <i class="fas fa-angle-down"></i></a></legend>
 
 <div class="content-section collapse" id="conveniosEnProduccion">
   <!-- Select Convenio -->
@@ -100,10 +100,10 @@
 </div>
 
 <!-- OTROS CONVENIOS -->
-<legend class="border-bottom mb-3">Otros convenios
+<legend class="border-bottom mb-3">
   <a style="text-decoration: none; color: #000;" data-bs-toggle="collapse"
   href="#otrosConvenios" aria-expanded="false" aria-controls="otrosConvenios">
-  <i class="fas fa-angle-down"></i></a></legend>
+  Otros convenios <i class="fas fa-angle-down"></i></a></legend>
 <div class="content-section collapse" id="otrosConvenios">
   <!-- Select Convenio -->
   <select class="form-control form-control-lg form-select mb-5" id="selectOtrosConvenios">
