+++ convenios_app/templates/informes/convenios_en_produccion.html
@@ -54,11 +54,11 @@ que ingresaron a partir del 01 de enero del 2020.</small>
   <div id='chartTareas'></div>
 </div>
 
-<!-- Días por etapa por años -->
+<!-- Días por etapa por años
 <legend class='border-bottom mb-2'>Días en etapas por años</legend>
 <div  class="mb-5">
   <div id='chartAnual'></div>
-</div>
+</div>  -->
 
 <!-- Listado convenios en producción -->
 <legend class='border-bottom mb-2'>Listado convenios en producción</legend>
