+++ convenios_app/templates/informes/otros_convenios.html
@@ -1,0 +1,10 @@
+{% extends 'layout.html' %}
+{% block content %}
+{{ JSGlue.include() }}
+<script src='{{ url_for("static", filename="script.js") }}'></script>
+<script type="text/javascript">
+  changeActive('#otrosConvenios')
+</script>
+
+HOLANDA QUE TALCA
+{% endblock content %}
