+++ convenios_app/templates/main/home.html
@@ -14,8 +14,8 @@
         Convenios firmados
       </div>
       <div class="col-lg-3">
-        <h1>{{ data.publicados }} <i class="fa-solid fa-certificate fa-fw fa-sm"></i></h1>
-        Convenios publicados
+        <h1>{{ data.por_firmar }} <i class="fa-solid fa-file"></i></h1>
+        Convenios por firmar
       </div>
       <div class="col-lg-3">
         <h1>{{ data.en_produccion }} <i class="fa-solid fa-right-left fa-fw fa-sm"></i></h1>
@@ -28,21 +28,70 @@
     </div>
 
     <div class="row mt-5 text-center">
+      <div class="col-lg-12" id='flujoFirmados'>
+        <table class="table">
+          <thead class="">
+            <tr>
+              <th scope="col"></th>
+              <th scope="col">Stock inicial</th>
+              <th scope="col">Nuevos</th>
+              <th scope="col">Firmados</th>
+              <th scope="col">Eliminados</th>
+              <th scope="col">Stock final</th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr>
+             <th scope="row">Enero</th>
+             <td>28</td>
+             <td>2</td>
+             <td>1</td>
+             <td>0</td>
+             <th>29</th>
+           </tr>
+           <tr>
+            <th scope="row">Febrero</th>
+            <td>29</td>
+            <td>0</td>
+            <td>2</td>
+            <td>0</td>
+            <th>27</th>
+          </tr>
+          <tr>
+           <th scope="row">Marzo</th>
+           <td>27</td>
+           <td>3</td>
+           <td>1</td>
+           <td>3</td>
+           <th>26</th>
+         </tr>
+         <tr>
+          <th scope="row">Abril</th>
+          <td>26</td>
+          <td>0</td>
+          <td>0</td>
+          <td>0</td>
+          <th>26</th>
+        </tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
+
+    <div class="row mt-5 text-center">
         <div class="col-lg-7" id="chartFirmados"></div>
         <div class="col-lg-5" id='chartEtapas'>
       </div>
     </div>
     <div class="row mt-5 text-center">
-      <div class="col-lg-12" id='chartFirmadosTotal'>
-
-      </div>
+      <div class="col-lg-12" id='chartFirmadosTotal'></div>
     </div>
   </div>
 
 
   <script type="text/javascript">
     // Load the Visualization API and the corechart package.
-    google.charts.load('current', {'packages':['corechart']});
+    google.charts.load('current', {'packages':['corechart', 'table']});
 
     // Set a callback to run when the Google Visualization API is loaded.
     google.charts.setOnLoadCallback(drawChart);
@@ -102,6 +151,7 @@
       var chartFirmados = new google.visualization.ColumnChart(document.getElementById('chartFirmados'));
       chartFirmados.draw(dataFirmados, optionsFirmados);
 
+
       // Gráfico convenios firmados total
       var data_total = {{ data_años_total | tojson }}
       data_total[0].push({role: 'style'});
@@ -125,12 +175,6 @@
       var chartFirmadosTotal = new google.visualization.ColumnChart(document.getElementById('chartFirmadosTotal'));
       chartFirmadosTotal.draw(dataFirmadosTotal, optionsFirmadosTotal);
      };
-
-
-
-
-
-
-     </script>
+  </script>
 
 {% endblock content %}
